Refactor the paths of the config files and have better referencing. Also fix an issue with configurable_attributes not showing whats expected.
