+++ lib/quandl/config/configurable.rb
@@ -1,7 +1,9 @@
+require 'active_support/inflector'
+
 module Quandl
   module Configurable
     def configuration
-      @configuration ||= Config.new(file_name, configuration_options)
+      @configuration ||= Quandl::Config.new(file_name, configuration_options)
     end
 
     private
